+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-Unreleased
+2.4.0
 ------
+## Attention: Update process can take some time!
 - CO-923 Fixed CrossSelling deletes in Shopware
 - CO-949 Removed payment trigger and payment table
 - CO-974 Changed logic when payment will be imported